pyproject.toml
[project]
name = "agentci"
version = "0.1.0"
description = "Continuous Integration for AI Agents. Catch cost spikes and logic regressions before production."
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
authors = [{name = "Sunil Closure"}]
keywords = ["ai", "agents", "testing", "llm", "ci-cd", "regression"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Testing",
    "Framework :: Pytest",
]

dependencies = [
    "pydantic>=2.0",
    "click>=8.0",
    "rich>=13.0",
    "pyyaml>=6.0",
    "opentelemetry-api>=1.20",
    "opentelemetry-sdk>=1.20",
]

[project.optional-dependencies]
openai = ["openai>=1.0"]
anthropic = ["anthropic>=0.20"]
langgraph = ["langgraph>=0.1"]
all = ["agentci[openai,anthropic,langgraph]"]
dev = ["pytest>=8.0", "ruff", "mypy"]

[project.scripts]
agentci = "agentci.cli:cli"

[project.entry-points.pytest11]
agentci = "agentci.pytest_plugin"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agentci"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.pyright]
include = ["src"]
extraPaths = ["src"]
